requirements.txt
# Production dependencies
Flask~=3.1.0
requests~=2.32.0
python-dotenv~=1.2.0

# Testing dependencies
pytest~=8.3.0
pytest-mock~=3.14.0
pytest-cov~=6.0.0
pytest-flask~=1.3.0
